All endpoints use the JaiaH5FileSet class Use mutex lock before accessing disk files, so only one python pseudo-thread starts the goby_log_tool process
