--- src/web/jdv/server/jaiabot_log_converter.service.template
@@ -1,10 +1,0 @@
-[Unit]
-Description=jaiabot_data_vision_web
-
-[Service]
-User=${user}
-WorkingDirectory=${working_directory}
-ExecStart=${working_directory}/jaiabot_log_converter.py -d /var/log/jaiabot/bot_offload -s 100000000000 -l INFO
-
-[Install]
-WantedBy=multi-user.target
